refactor: split up tests
split up tests in their own files
